requirements.txt
Django==1.10.5
django-pagedown==0.1.1
Markdown==2.6.7
